Update and rename new_start to list_agumentation
I'm trying to find out why, when I want to re print 'invite_msg1' after popping out the old value from the list, that being brook. That even when calling, guest_list1 in a print. It  has the change (which is kyle). But brook still shows up when re-running the 'invite_msg1'. 

I feel i shouldn't have to redo a entire list, to see the change of kyle from brooks, when the pop() method has done what I needed it to do. I feel I should be able to print invite_list1, an the elements inside should take on the new value. 

Thank in advance